fix: init chat
