+++ projects/app/src/components/ChatBox/utils.ts
@@ -1,6 +1,5 @@
 import { SystemInputEnum } from '@/constants/app';
 import { FlowModuleTypeEnum } from '@/constants/flow';
-import { chatModelList } from '@/web/common/store/static';
 import { AppModuleItemType, VariableItemType } from '@/types/app';
 
 export const getGuideModule = (modules: AppModuleItemType[]) =>
@@ -22,13 +21,4 @@ export const splitGuideModule = (guideModules?: AppModuleItemType) => {
     variableModules,
     questionGuide
   };
-};
-export const getChatModelNameList = (modules: AppModuleItemType[]): string[] => {
-  const chatModules = modules.filter((item) => item.flowType === FlowModuleTypeEnum.chatNode);
-  return chatModules
-    .map((item) => {
-      const model = item.inputs.find((input) => input.key === 'model')?.value;
-      return chatModelList.find((item) => item.model === model)?.name || '';
-    })
-    .filter((item) => item);
 };
